requirements.txt
asgiref==3.9.1
boto3==1.40.35
botocore==1.40.35
Django==5.2.6
django-storages==1.14.6
djangorestframework==3.16.1
python-dotenv==1.0.1
jmespath==1.0.1
pip==25.2
psycopg2==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
